requirements.txt
asgiref==3.8.1
beautifulsoup4==4.12.3
bs4==0.0.2
certifi==2024.7.4
charset-normalizer==3.3.2
Django==5.1
django-cors-headers==4.4.0
django-environ==0.11.2
django-filter==24.3
djangorestframework==3.15.2
djangorestframework-jwt==1.11.0
djangorestframework-simplejwt==5.3.1
idna==3.7
Markdown==3.6
polygon-api-client==1.14.2
PyJWT==1.7.1
requests==2.32.3
soupsieve==2.6
sqlparse==0.5.1
tzdata==2024.1
urllib3==2.2.2
websockets==12.0
